Uart refactor:
3. Switch case instead of if else chain

2. Split off on cmd functions to command tree

1. Uart shell split to files in subfolder
