+++ BrushedMotorControllerFirmware/include/uart_shell/shell_utils.h
@@ -1,0 +1,11 @@
+/* SPDX-License-Identifier: Apache-2.0
+ *
+ * Copyright (c) 2024 Jakub Mazur
+ *
+ */
+
+#if defined(CONFIG_UART_SHELL_SUPPORT)
+#include "driver.h"
+
+enum ChannelNumber get_relevant_channel(void);
+#endif
